+++ src/main/java/org/lagonette/hellos/bean/StatusPaymentEnum.java
@@ -3,6 +3,7 @@ package org.lagonette.hellos.bean;
 public enum StatusPaymentEnum {
     todo("À faire"),
     tooHigh("Montant trop haut"),
+    previewOK("Reactiver les paiements pour effectuer le paiement"),
     success("Succès"),
     fail("Echec");