+++ src/main/java/org/lagonette/hellos/service/PaymentService.java
@@ -6,7 +6,9 @@ import org.lagonette.hellos.bean.Notification;
 import org.lagonette.hellos.bean.ProcessResult;
 import org.lagonette.hellos.bean.StatusPaymentEnum;
 import org.lagonette.hellos.bean.helloasso.notification.HelloAssoPaymentNotification;
+import org.lagonette.hellos.entity.Configuration;
 import org.lagonette.hellos.entity.Payment;
+import org.lagonette.hellos.repository.ConfigurationRepository;
 import org.lagonette.hellos.repository.PaymentRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,18 +17,22 @@ import org.springframework.stereotype.Component;
 import java.io.IOException;
 import java.util.Map;
 
+import static org.lagonette.hellos.service.ConfigurationService.MAIL_RECIPIENT;
+import static org.lagonette.hellos.service.ConfigurationService.PAYMENT_AUTOMATIC_ENABLED;
+
 @Component
 public class PaymentService {
     private final Logger LOGGER = LoggerFactory.getLogger(this.getClass());
-    private final String ERROR_SUBJECT = "[Hellos] Erreur lors du traitement";
 
+    private final ConfigurationRepository configurationRepository;
     private final PaymentRepository paymentRepository;
     private final MailService mailService;
     private final Dotenv dotenv;
     private final HelloAssoService helloAssoService;
     private final CyclosService cyclosService;
 
-    public PaymentService(PaymentRepository paymentRepository, MailService mailService, Dotenv dotenv, HelloAssoService helloAssoService, CyclosService cyclosService) {
+    public PaymentService(ConfigurationRepository configurationRepository, PaymentRepository paymentRepository, MailService mailService, Dotenv dotenv, HelloAssoService helloAssoService, CyclosService cyclosService) {
+        this.configurationRepository = configurationRepository;
         this.paymentRepository = paymentRepository;
         this.mailService = mailService;
         this.dotenv = dotenv;
@@ -37,7 +43,7 @@ public class PaymentService {
     public ProcessResult handleNewPayment(ProcessResult processResult, String helloAssoPaymentNotificationWrapper) throws IOException {
 
         if (helloAssoPaymentNotificationWrapper == null) {
-            LOGGER.error("The payment is empty : {}");
+            LOGGER.error("The payment is empty");
             processResult.getErrors().add("Le paiement est vide ");
             processResult.setStatusPayment(StatusPaymentEnum.fail);
             return processResult;
@@ -47,7 +53,7 @@ public class PaymentService {
 
         // check if payment is real
         Map<Boolean, Notification> validPayment = helloAssoService.isValidPayment(helloAssoPaymentNotification);
-        if (validPayment.keySet().contains(false)) {
+        if (validPayment.containsKey(false)) {
             LOGGER.debug("The payment is not valid : {}", helloAssoPaymentNotificationWrapper);
             processResult.getErrors().add("Le paiement n'est pas valide :" + helloAssoPaymentNotificationWrapper);
             processResult.setStatusPayment(StatusPaymentEnum.fail);
@@ -55,40 +61,60 @@ public class PaymentService {
         }
         Notification notification = validPayment.get(true);
         final Payment paymentInDatabase = paymentRepository.findById(notification.getId());
+
         if (paymentInDatabase == null) {
             // register payment in database (convert cent to euro)
             Payment payment = new Payment(notification.getId(), notification.getDate(), (float) notification.getAmount() / 100,
                     notification.getFirstName(), notification.getName(), notification.getEmail());
-            paymentRepository.save(payment);
+            Payment paymentSaved = paymentRepository.save(payment);
+
+            if ("true".equals(configurationRepository.findById(PAYMENT_AUTOMATIC_ENABLED).orElse(new Configuration(PAYMENT_AUTOMATIC_ENABLED, "false")).getValue())) {
+                LOGGER.info("automatic payment to be proceed");
+                processResult = creditCyclosAccount(processResult, paymentSaved);
+                if (StatusPaymentEnum.success.equals(processResult.getStatusPayment())) {
+                    mailService.sendEmail(dotenv.get("MAIL_RECIPIENT"), "[Hellos] Paiement réussi :)",
+                            "Un paiement a été effectué avec succès.\nId : " + paymentSaved.getId());
+                } else {
+                    mailService.sendEmail(dotenv.get("MAIL_RECIPIENT"), "[Hellos] Paiement en échec :(",
+                            "Un paiement n'a pas pu être effectué.\nId : " + paymentSaved.getId());
+                }
+            }
         } else {
             LOGGER.debug("Payment already inserted in database : {}", notification.getId());
         }
+
         return processResult;
     }
 
-    public void creditAccount(ProcessResult processResult, int paymentId) {
+    public ProcessResult creditAccount(ProcessResult processResult, int paymentId) {
         Payment payment = paymentRepository.findById(paymentId);
         if (payment == null) {
             LOGGER.error("Payment not found : {}", paymentId);
-            return;
+            return processResult;
         }
 
-        processResult = cyclosService.creditAccount(processResult, paymentId);
+        creditCyclosAccount(processResult, payment);
+        return processResult;
+    }
+
+    private ProcessResult creditCyclosAccount(ProcessResult processResult, Payment payment) {
+        processResult = cyclosService.creditAccount(processResult, payment.getId());
         payment.setStatus(processResult.getStatusPayment());
         if (!processResult.getErrors().isEmpty() && processResult.getErrors().toString().length() > Payment.ERROR_LENGTH) {
             payment.setError(processResult.getErrors().toString().substring(0, Payment.ERROR_LENGTH - 100));
-        }
-        else {
+        } else {
             payment.setError(processResult.getErrors().toString());
         }
         paymentRepository.save(payment);
-        sendErrorEmail(processResult, paymentId);
+        sendErrorEmail(processResult, payment.getId());
+        return processResult;
     }
 
     public void sendErrorEmail(ProcessResult processResult, int paymentId) {
         if (!StatusPaymentEnum.success.equals(processResult.getStatusPayment())) {
             String body = "Liste des erreurs pour le paiement " + paymentId + ": \n " + processResult.getErrors().toString();
-            mailService.sendEmail(dotenv.get("MAIL_RECIPIENT"), ERROR_SUBJECT, body);
+            String ERROR_SUBJECT = "[Hellos] Erreur lors du traitement";
+            mailService.sendEmail(configurationRepository.findById(MAIL_RECIPIENT).get().getValue(), ERROR_SUBJECT, body);
         }
     }
 }