+++ src/test/java/org/lagonette/hellos/service/CyclosServiceTest.java
@@ -16,7 +16,9 @@ import org.lagonette.hellos.bean.cyclos.CyclosGroup;
 import org.lagonette.hellos.bean.cyclos.CyclosPerformPaymentResponse;
 import org.lagonette.hellos.bean.cyclos.CyclosTransaction;
 import org.lagonette.hellos.bean.cyclos.CyclosUser;
+import org.lagonette.hellos.entity.Configuration;
 import org.lagonette.hellos.entity.Payment;
+import org.lagonette.hellos.repository.ConfigurationRepository;
 import org.lagonette.hellos.repository.PaymentRepository;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -26,8 +28,10 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.lagonette.hellos.service.ConfigurationService.PAYMENT_CYCLOS_ENABLED;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -40,6 +44,9 @@ class CyclosServiceTest {
 
     @Mock
     private MailService mailService;
+
+    @Mock
+    private ConfigurationRepository configurationRepository;
 
     @Mock
     private PaymentRepository paymentRepository;
@@ -67,7 +74,7 @@ class CyclosServiceTest {
     }
 
     @Test
-    void creditAccount_shouldBeSuccessfull() throws JsonProcessingException {
+    void creditAccount_shouldBeSuccessful() throws JsonProcessingException {
         // GIVEN
         ObjectMapper objectMapper = new ObjectMapper();
         CyclosUser cyclosUser = new CyclosUser();
@@ -100,6 +107,8 @@ class CyclosServiceTest {
         when(dotenv.get("CYCLOS_URL")).thenReturn(baseUrl);
         when(dotenv.get("CYCLOS_USER")).thenReturn("user");
         when(dotenv.get("CYCLOS_PWD")).thenReturn("pwd");
+
+        when(configurationRepository.findById(PAYMENT_CYCLOS_ENABLED)).thenReturn(Optional.of(new Configuration(PAYMENT_CYCLOS_ENABLED, "true")));
         ProcessResult processResult = new ProcessResult();
 
         // WHEN
@@ -107,6 +116,52 @@ class CyclosServiceTest {
 
         // THEN
         assertThat(finalProcessResult.getStatusPayment()).isEqualTo(StatusPaymentEnum.success);
+        assertThat(finalProcessResult.getErrors()).isEmpty();
+    }
+
+    @Test
+    void creditAccount_shouldBePreviewed() throws JsonProcessingException {
+        // GIVEN
+        ObjectMapper objectMapper = new ObjectMapper();
+        CyclosUser cyclosUser = new CyclosUser();
+        CyclosGroup group = new CyclosGroup();
+        group.setInternalName("group-name");
+        cyclosUser.setGroup(group);
+
+        // Mock all HTTP calls
+        // first call, get user
+        mockBackEnd.enqueue(new MockResponse()
+                .setBody(objectMapper.writeValueAsString(cyclosUser))
+                .addHeader("Content-Type", "application/json"));
+        // second call, get previous transactions
+        List<CyclosTransaction> transactions = new ArrayList<>();
+        mockBackEnd.enqueue(new MockResponse()
+                .setBody(objectMapper.writeValueAsString(transactions))
+                .addHeader("Content-Type", "application/json"));
+        // third call, post payment
+        CyclosPerformPaymentResponse paymentDone = new CyclosPerformPaymentResponse();
+        List<CyclosPerformPaymentResponse> cyclosPerformPaymentResponse = Collections.singletonList(paymentDone);
+        mockBackEnd.enqueue(new MockResponse()
+                .setBody(objectMapper.writeValueAsString(cyclosPerformPaymentResponse))
+                .addHeader("Content-Type", "application/json"));
+        payment = Payment.PaymentBuilder.aPayment().build();
+        when(paymentRepository.findById(1)).thenReturn(payment);
+        when(dotenv.get("CYCLOS_EMISSION_PART_INTERNAL")).thenReturn("emission");
+        when(dotenv.get("CYCLOS_GROUP_PART_INTERNAL")).thenReturn("group-name");
+        when(dotenv.get("CYCLOS_GROUP_PRO_INTERNAL")).thenReturn("pro-group-name");
+
+        when(dotenv.get("CYCLOS_URL")).thenReturn(baseUrl);
+        when(dotenv.get("CYCLOS_USER")).thenReturn("user");
+        when(dotenv.get("CYCLOS_PWD")).thenReturn("pwd");
+
+        when(configurationRepository.findById(PAYMENT_CYCLOS_ENABLED)).thenReturn(Optional.of(new Configuration(PAYMENT_CYCLOS_ENABLED, "false")));
+        ProcessResult processResult = new ProcessResult();
+
+        // WHEN
+        final ProcessResult finalProcessResult = cyclosService.creditAccount(processResult, 1);
+
+        // THEN
+        assertThat(finalProcessResult.getStatusPayment()).isEqualTo(StatusPaymentEnum.previewOK);
         assertThat(finalProcessResult.getErrors()).isEmpty();
     }